images now work in docker
