+++ .gitignore
@@ -5,3 +5,6 @@ report*
 /mongo*
 .idea
 .DS_Store
+img/*
+!/img/404.png
+!/img/logo.png
