+++ site/nginx/nginx.conf
@@ -2,9 +2,14 @@ server {
  listen 80;
  server_name frontend;
  location / {
-   # This would be the directory where your React app's static files are stored at
    root /usr/share/nginx/html;
    try_files $uri /index.html;
+ }
+ location /img/ {
+   root /var/img;
+   rewrite /img/(.*) /$1 break; 
+   try_files $uri /404.png;
+  autoindex on;
  }
  location /admin/api {
    rewrite /admin/api/(.*) /admin/$1 break;
